requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.2.0
coverage==7.2.7
Django==4.2.3
django-pdb==0.6.2
djangorestframework==3.14.0
docopt==0.6.2
factory-boy==3.2.1
Faker==18.13.0
idna==3.4
pipreqs==0.4.13
psycopg2-binary==2.9.6
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
urllib3==2.0.3
yarg==0.1.9
